simplify API by removing errors and using panics instead
Previously, when using an invalid mode or an unsupported hashing algorithm, an error would be returned. However, both of these problems are programmer induced issues: an invalid mode occurs when a new mode is added but is not handled in the params() method, and an unsupported hashing algorithm error is the result of a user specifying it in the Opts struct. These cases are better handled as panics.